backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib>=1.3.2

twilio
sendgrid
sqlalchemy
apscheduler
shap
python-dotenv
feedparser>=6.0.11
